% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_coordinates.btn.R
\name{cell_coordinates.btn}
\alias{cell_coordinates.btn}
\title{Get cell x, y and z coordinates from a btn object}
\usage{
\method{cell_coordinates}{btn}(btn, include_faces = FALSE)
}
\arguments{
\item{btn}{btn object}

\item{include_faces}{logical; should face coordinates be included?}
}
\value{
list with cell coordinate 3d arrays
}
\description{
Get cell x, y and z coordinates from a btn object
}

