% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_btn_to_dis.R
\name{convert_btn_to_dis}
\alias{convert_btn_to_dis}
\title{Convert RMT3DMS btn to RMODFLOW dis object}
\usage{
convert_btn_to_dis(btn)
}
\arguments{
\item{btn}{btn object}
}
\value{
dis object
}
\description{
Convert RMT3DMS btn to RMODFLOW dis object
}

