% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_btn_to_real.R
\name{convert_btn_to_real}
\alias{convert_btn_to_real}
\title{Convert mt3dms coordinates to real world coordinates}
\usage{
convert_btn_to_real(x = NULL, y = NULL, z = NULL, i = NULL, j = NULL,
  k = NULL, prj, btn = NULL)
}
\arguments{
\item{x}{mt3dms x coordinate}

\item{y}{mt3dms y coordinate}

\item{z}{mt3dms z coordinate}

\item{i}{mt3dms row number}

\item{j}{mt3dms column number}

\item{k}{mt3dms layer number}

\item{prj}{prj object}

\item{btn}{btn object}
}
\value{
data frame with real world x and y coordinates
}
\description{
Convert mt3dms coordinates to real world coordinates
}
\details{
Provide either xyz or ijk
}

