% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert_real_to_btn}
\alias{convert_real_to_btn}
\title{Convert real world coordinates to mt3dms coordinates}
\usage{
convert_real_to_btn(x, y, prj, z = NULL, btn = NULL)
}
\description{
Convert real world coordinates to mt3dms coordinates
}

