% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.mt3dms_3d_array}
\alias{plot.mt3dms_3d_array}
\title{Plot a layer of a MT3DMS 3D array}
\usage{
\method{plot}{mt3dms_3d_array}(mt3dms_3d_array, i = NULL, j = NULL,
  k = NULL, btn, mask = btn$ICBUND,
  zlim = range(mt3dms_3d_array[ifelse0(is.null(i),
  c(1:dim(mt3dms_3d_array)[1]), i), ifelse0(is.null(j),
  c(1:dim(mt3dms_3d_array)[2]), j), ifelse0(is.null(k),
  c(1:dim(mt3dms_3d_array)[3]), k)], finite = TRUE),
  color.palette = rev_rainbow, nlevels = 7, ...)
}
\arguments{
\item{mt3dms_3d_array}{An object of class mt3dms_3d_array, or a 3D array}

\item{mask}{A 3D icbund array with 1 or TRUE indicating active cells, and 0 or F indicating inactive cells}

\item{zlim}{}

\item{color.palette}{A color palette for imaging the parameter values}

\item{nlevels}{}

\item{layer}{The number of the layer to plot}

\item{levels}{}

\item{main}{}
}
\value{
None
}
\description{
\code{plot.mt3dms_3d_array} plots a 2D section through a MT3DMS 3D array.
}

