% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rmt3dms_2d_array.R
\name{plot.rmt3dms_2d_array}
\alias{plot.rmt3dms_2d_array}
\title{Plot a MODFLOW 2D array}
\usage{
\method{plot}{rmt3dms_2d_array}(rmt3dms_2d_array, btn, mask = {    
  warning("Using first icbund layer as mask.", call. = FALSE)     btn$icbund[, ,
  1] }, ...)
}
\arguments{
\item{rmt3dms_2d_array}{an object of class rmt3dms_2d_array, or a 2D matrix}

\item{btn}{basic transport file object}

\item{...}{arguments passed to plot.rmodflow_2d_array}
}
\value{
ggplot2 object or layer; if plot3D is TRUE, nothing is returned and the plot is made directly
}
\description{
\code{plot.rmt3dms_2d_array} plots a MODFLOW 2D array.
}

