% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rmt3dms_3d_array.R
\name{plot.rmt3dms_3d_array}
\alias{plot.rmt3dms_3d_array}
\title{Plot a layer of a MT3DMS 3D array}
\usage{
\method{plot}{rmt3dms_3d_array}(rmt3dms_3d_array, i = NULL, j = NULL,
  k = NULL, btn, mask = btn$icbund, ...)
}
\arguments{
\item{rmt3dms_3d_array}{an object of class rmt3dms_3d_array, or a 3D array}

\item{i}{row number to plot}

\item{j}{column number to plot}

\item{k}{layer number to plot}

\item{btn}{basic transport file object}

\item{mask}{a 3D array with 0 or F indicating inactive cells; defaults to btn$icbund}

\item{...}{arguments provided to RMODFLOW::plot.rmodflow_3d_array}
}
\value{
ggplot2 object or layer; if plot3D is TRUE, nothing is returned and the plot is made directly
}
\description{
\code{plot.rmt3dms_3d_array} plots a 2D section through a MT3DMS 3D array.
}

