% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.adv}
\alias{read.adv}
\title{Read an MT3DMS advection package file}
\usage{
read.adv(file)
}
\arguments{
\item{file}{Filename; typically *.adv}
}
\value{
Object of class adv
}
\description{
\code{read.adv} reads in an MT3DMS advection package file and returns it as an \code{\link{RMT3DMS}} adv object.
}

