% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.prj}
\alias{read.prj}
\title{Read a projection file}
\usage{
read.prj(file)
}
\arguments{
\item{file}{Filename; typically *.prj}
}
\value{
Object of class prj
}
\description{
\code{read.prj} reads in projection file and returns it as an \code{\link{RMT3DMS}} prj object.
}

