% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.rct}
\alias{read.rct}
\title{Read an MT3DMS chemical reaction package file}
\usage{
read.rct(file, btn)
}
\arguments{
\item{file}{Filename; typically *.rct}
}
\value{
Object of class rct
}
\description{
\code{read.rct} reads in an MT3DMS dispersion package file and returns it as an \code{\link{RMT3DMS}} rct object.
}

