% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_btn}
\alias{read_btn}
\title{Read an MT3DMS basic transport package file}
\usage{
read_btn(file)
}
\arguments{
\item{file}{Filename; typically *.btn}
}
\value{
Object of class btn
}
\description{
\code{read.btn} reads in an MT3DMS basic transport package file and returns it as an \code{\link{RMT3DMS}} btn object.
}

