% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dsp.R
\name{read_dsp}
\alias{read_dsp}
\title{Read an MT3DMS dispersion package file}
\usage{
read_dsp(file, btn)
}
\arguments{
\item{file}{filename; typically '*.dsp'}

\item{btn}{basic transport package file object}
}
\value{
object of class dsp
}
\description{
\code{read_dsp} reads in an MT3DMS dispersion package file and returns it as an \code{\link{RMT3DMS}} dsp object.
}

