% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_gcg.R
\name{read_gcg}
\alias{read_gcg}
\title{Read an MT3DMS generalized conjugate gradient solver package file}
\usage{
read_gcg(file)
}
\arguments{
\item{file}{filename; typically '*.gcg'}
}
\value{
object of class gcg
}
\description{
\code{read_gcg} reads in an MT3DMS generalized conjugate gradient solver package file and returns it as an \code{\link{RMT3DMS}} gcg object.
}

