% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_mas}
\alias{read_mas}
\title{Read an MT3DMS mass balance summary file}
\usage{
read_mas(file)
}
\arguments{
\item{file}{Filename; typically *.mas}
}
\value{
Object of class mas
}
\description{
\code{read.mas} reads in an MT3DMS mass balance summary file and returns it as an \code{\link{RMT3DMS}} mas object.
}

