% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{read_mt3dms_variables}
\alias{read_mt3dms_variables}
\title{Read mt3dms variables
If all are numbers, returns numeric, otherwise returns character vector}
\usage{
read_mt3dms_variables(remaining_lines)
}
\description{
Read mt3dms variables
If all are numbers, returns numeric, otherwise returns character vector
}

