% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_rct.R
\name{read_rct}
\alias{read_rct}
\title{Read an MT3DMS chemical reaction package file}
\usage{
read_rct(file, btn)
}
\arguments{
\item{file}{filename; typically '*.rct'}

\item{btn}{basic transport package file object}
}
\value{
object of class rct
}
\description{
\code{read_rct} reads in an MT3DMS dispersion package file and returns it as an \code{\link{RMT3DMS}} rct object.
}

