% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_ssm}
\alias{read_ssm}
\title{Read an MT3DMS sink & source mixing package file}
\usage{
read_ssm(file, btn)
}
\arguments{
\item{file}{Filename; typically *.ssm}
}
\value{
Object of class ssm
}
\description{
\code{read.ssm} reads in an MT3DMS sink & source mixing package file and returns it as an \code{\link{RMT3DMS}} ssm object.
}

