% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_ucn}
\alias{read_ucn}
\title{Read an MT3DMS unformatted concentration file}
\usage{
read_ucn(file, btn)
}
\arguments{
\item{file}{Filename; typically *.ucn}
}
\value{
Object of class ucn
}
\description{
\code{read.ucn} reads in an MT3DMS unformatted concentration file and returns it as an \code{\link{RMT3DMS}} ucn object.
}

