% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-tob.R
\name{rmf_create_tob}
\alias{rmf_create_tob}
\title{Create an \code{RMT3DMS} tob object}
\usage{
rmf_create_tob(
  locations,
  time_series,
  btn,
  hydraulic_conductivity = array(1, dim = c(dis$nrow, dis$ncol, dis$nlay)),
  prj = NULL,
  cscale = 1,
  ioutcobs = 1,
  iconclog = 0,
  iconcintp = 1,
  outnam = "output",
  save_ocn = 665,
  save_mfx = 0,
  save_pst = 0,
  unique_obsnam = FALSE
)
}
\arguments{
\item{locations}{data.frame or \code{sf POINT} object with the name, x and y coordinates (if data.frame), and top and bottom of the monitoring well filter}

\item{time_series}{data.frame with the name of the monitoring well filter and the observation time and concentration column(s) (css1, css2 etc..)}

\item{btn}{RMT3DMS btn object}

\item{prj}{RMODFLOW prj object; provide if coordinates in \code{locations} are real world coordinates}

\item{unique_obsnam}{logical; should an ID number be added to obsnam for filters with observations at different times? Defaults to FALSE.}

\item{iutobsv}{file unit (> 0) for saving observation data in a file; the file unit should be included as DATA type in the name file; defaults to 665}

\item{tobdry}{value of the simulated equivalent for a dry cell used in the observation output file}

\item{noprint}{logical; if TRUE, the input and output are not printed in the listing file}

\item{tomult}{time offset multiplier, to convert the time unit in \code{time_series} to the MODFLOW time unit}

\item{itt}{integer flag, either a single value or a value for every observation well as defined in \code{locations}. Identifies whether hydraulic heads (1; default) or changes in hydraulic heads (2) are used as observations. See details.}
}
\value{
Object of class tob
}
\description{
\code{rmf_create_tob} creates an RMT3DMS tob object
}
\details{
Regardless of the value of \code{itt}, observed heads should be specified in \code{time_series}. MODFLOW will calculate the changes in hydraulic head internally.
The \code{locations} data.frame must have columns named \code{name, x, y, top, bottom}.
The \code{time_series} data.frame must have columns named \code{name, time, concentration}.
}
\seealso{
\code{\link{rmf_read_tob}}, \code{\link{rmf_write_tob}}
}
