% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-tob.R
\name{rmf_read_tob}
\alias{rmf_read_tob}
\alias{rmf_read_ob_tob}
\title{Read a RMT3D%S Transport Observation Package file}
\usage{
rmf_read_tob(
  file = {     cat("Please select tob file ...\\n")     file.choose() },
  ...
)

rmf_read_ob_tob(...)
}
\arguments{
\item{file}{filename; typically '*.tob'}

\item{...}{arguments passed to \code{rmfi_parse_variables}. Can be ignored when input is 'free' format.}
}
\value{
object of class tob
}
\description{
\code{rmt_read_tob} reads in a MODFLOW head observations file and returns it as an \code{\link{RMODFLOW}} tob object.
}
\section{Functions}{
\itemize{
\item \code{rmf_read_ob_tob}: Compatible with default ModelMuse file extensions
}}

\seealso{
\code{\link{rmf_create_tob}}, \code{\link{rmf_write_tob}} and \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?tob.htm}
}
