% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-tob.R
\name{rmf_write_tob}
\alias{rmf_write_tob}
\title{Write a MODFLOW head observations file}
\usage{
rmf_write_tob(
  tob,
  file = {    
    cat("Please select tob file to overwrite or provide new filename ...\\n")    
    file.choose() },
  ...
)
}
\arguments{
\item{tob}{a \code{\link{RMODFLOW}} tob object}

\item{file}{filename to write to; typically '*.tob'}

\item{...}{arguments passed to \code{rmfi_write_variables} when writing a fixed format file.}
}
\value{
\code{NULL}
}
\description{
Write a MODFLOW head observations file
}
\seealso{
\code{\link{rmf_create_tob}}, \code{\link{rmf_read_tob}} and \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?tob.htm}
}
