% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmt_as_list}
\alias{rmt_as_list}
\alias{rmt_as_list.wel}
\alias{rmt_as_list.drn}
\alias{rmt_as_list.chd}
\alias{rmt_as_list.riv}
\alias{rmt_as_list.ghb}
\title{Generic to convert objects to rmt_list}
\usage{
rmt_as_list(...)

\method{rmt_as_list}{wel}(wel, cwel, kper, ...)

\method{rmt_as_list}{drn}(drn, cdrn, kper, ...)

\method{rmt_as_list}{chd}(chd, cchd, kper, ...)

\method{rmt_as_list}{riv}(riv, criv, kper, ...)

\method{rmt_as_list}{ghb}(ghb, cghb, kper, ...)
}
\arguments{
\item{...}{additional numeric vectors containing concentrations (one value for each feature) or column names. One vector/name for each species > 1 (in order).}

\item{wel}{\code{RMODFLOW} wel object}

\item{cwel}{concentrations in the wells for species 1. A value must be supplied for every well. Alternatively, a character can be entered referencing a column in the wel data set to use.}

\item{kper}{numeric vector with the stress period numbers during which the list is active.}

\item{drn}{\code{RMODFLOW} drn object}

\item{cdrn}{concentrations in the drains for species 1. A value must be supplied for every drain. Alternatively, a character can be entered referencing a column in the drn data set to use.}

\item{chd}{\code{RMODFLOW} chd object}

\item{cchd}{concentrations in the constant-head cells for species 1. A value must be supplied for every constant-head cell. Alternatively, a character can be entered referencing a column in the chd data set to use.}

\item{riv}{\code{RMODFLOW} riv object}

\item{criv}{concentrations in the river cells for species 1. A value must be supplied for every river cell. Alternatively, a character can be entered referencing a column in the riv data set to use.}

\item{ghb}{\code{RMODFLOW} ghb object}

\item{cghb}{concentrations in the general-head boundary cells for species 1. A value must be supplied for every general-head boundary cell. Alternatively, a character can be entered referencing a column in the ghb data set to use.}
}
\value{
a \code{rmt_list} object

a \code{rmt_list} object

a \code{rmt_list} object

a \code{rmt_list} object

a \code{rmt_list} object
}
\description{
Generic to convert objects to rmt_list
}
\details{
kper should be one of the stress periods defined in the flow model unless only one stress period is present in the flow model.
}
\examples{
library(RMODFLOW)
m <- rmf_read(rmf_example_file('example-model.nam'), verbose = FALSE)
rmt_as_list(m$wel, 6.5, kper = 1)

m$wel$data$chloride <- 6.5
rmt_as_list(m$wel, 'chloride', kper = 1)
# 3 species
cchd1 <- rnorm(nrow(m$chd$data), 70, 50)
rmt_as_list(m$chd, cchd1, kper = c(2, 3), cchd1 * 2, cchd1 / 2)

# stress periods should be the same when flow object has more than 1 stress period
\dontrun {
rmt_as_list(m$chd, cchd1, kper = c(2, 3))
}
}
