% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmt_convert_btn_to_dis}
\alias{rmt_convert_btn_to_dis}
\title{Convert RMT3DMS btn to RMODFLOW dis object}
\usage{
rmt_convert_btn_to_dis(
  btn,
  nper = btn$nper,
  perlen = btn$perlen,
  nstp = btn$nstp,
  tsmult = btn$tsmult,
  sstr = c("SS", rep("TR", nper - 1)),
  ...
)
}
\arguments{
\item{btn}{\code{RMT3DMS} btn object}

\item{nper}{number of stress periods in the dis object. Defaults to \code{btn$nper}. See details.}

\item{perlen}{vector of stress period lengths in the dis object. Defaults to \code{btn$perlen}. See details.}

\item{nstp}{vector of stress period time steps in the dis object. Defaults to \code{btn$nstp}. See details.}

\item{tsmult}{vector of stress period time step multipliers in the dis object. Defaults to \code{btn$tsmult}. See details.}

\item{sstr}{character vector with steady state ('SS') or transient ('TR') stress period indicator; defaults to first period 'SS' and all others 'TR'}

\item{...}{additional arguments passed to \code{RMODFLOW::rmf_create_dis}}
}
\value{
\code{RMODFLOW} dis object
}
\description{
Convert RMT3DMS btn to RMODFLOW dis object
}
\details{
Conversion works best when number of stress-periods is equal in the dis and btn objects.
 \code{nper}, \code{perlen}, \code{nstp} & \code{tsmult} should only differ from the defaults when the flow model is steady-state 
 with only 1 stress period. In that case, MT3DMS allows for differing stress periods between the flow and transport model.
}
\examples{
btn <- rmt_create_btn()
rmt_convert_btn_to_dis(btn)

# BTN with 2 stress periods but flow model is single stress period steady-state
btn <- rmt_create_btn(nper = 2, perlen = c(1000, 500))
rmt_convert_btn_to_dis(btn, nper = 1, perlen = 1, nstp = 1, tsmult = 1)

otherwise converts to dis object with only first stress period steady-state
rmt_convert_btn_to_dis(btn)

}
\seealso{
\code{\link{rmt_convert_dis_to_btn}}
}
