% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmt_convert_rmf_to_rmt}
\alias{rmt_convert_rmf_to_rmt}
\title{Convert a RMODFLOW object to a RMT3DMS object}
\usage{
rmt_convert_rmf_to_rmt(obj)
}
\arguments{
\item{obj}{either a \code{rmf_2d_array}, \code{rmf_3d_array}, \code{rmf_4d_array} or a \code{rmf_list}}
}
\value{
obj with class \code{rmt_2d_array}, \code{rmt_3d_array}, \code{rmt_4d_array} or \code{rmt_list}
}
\description{
\code{rmf_convert_rmf_to_rmt} adds the corresponding RMT3DMS class to an RMODFLOW object
}
