% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmt_convert_tob_to_locations}
\alias{rmt_convert_tob_to_locations}
\title{Convert a tob object to a locations data frame}
\usage{
rmt_convert_tob_to_locations(tob, btn, prj = RMODFLOW::rmf_get_prj(btn))
}
\arguments{
\item{btn}{\code{RMT3DMS} btn object}

\item{prj}{prj object}

\item{hob}{\code{RMT3DMS} tob object}
}
\value{
a data frame containing name, screened layer proportion pr, cell indices k, i, j and x, y & z coordinates
}
\description{
\code{rmt_convert_tob_to_locations} converts a tob object holding concentration observations to a locations data frame
}
\details{
returned z coordinate represents the center of the cell, not the top or bottom of the well screen. 
         The returned locations represent the concentration observations, not the mass flux observations.
}
\seealso{
\code{\link{rmt_create_tob}}
}
