% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt3dms.R
\name{rmt_create}
\alias{rmt_create}
\title{Create an \code{RMT3DMS} mt3dms object}
\usage{
rmt_create(..., recreate_nam = FALSE, ftl, ftl_free = NULL, basename = NULL)
}
\arguments{
\item{...}{(list of) \code{RMT3DMS} objects of class \code{rmt_package} to be included in the mt3dms object. If a \code{nam} object is not provided, it is added automatically.}

\item{recreate_nam}{logical; if a \code{nam} object is supplied, should it be recreated from scratch ? Defaults to FALSE}

\item{ftl}{path to the flow-transport link file in the \code{nam} object; typically '.ftl'. Only used when a \code{nam} object is not supplied or when \code{recreate_nam = TRUE}.}

\item{ftl_free}{logical; is the flow-transport link file in the \code{nam} object written in free (formatted) format (TRUE) or binary (unformatted) (FALSE)? if NULL (default), it is guessed from reading \code{ftl}. Only used when a \code{nam} object is not supplied or when \code{recreate_nam = TRUE}.}

\item{basename}{character specifying the basename of the files if the nam object is (re)created. The default (\code{NULL}) sets input basenames to 'input' and output to 'output'.}
}
\value{
a \code{mt3dms} object which is a list containing all MT3DMS/MT3D-USGS packages
}
\description{
\code{rmt_create} creates an \code{RMT3DMS} mt3dms object from \code{rmt_package} objects
}
\examples{
btn <- rmt_create_btn()
adv <- rmt_create_adv()
gcg <- rmt_create_gcg()
nam <- rmt_create_nam(btn, adv, gcg, ftl = 'output.ftl', ftl_free = TRUE)

rmt_create(btn, adv, gcg, nam)
rmt_create(btn, adv, gcg, nam, recreate_nam = TRUE, basename = 'ex1')
rmt_create(btn, adv, gcg, ftl = 'output.ftl', ftl_free = FALSE)
}
\seealso{
\code{\link{rmt_read}}, \code{\link{rmt_write}}
}
