% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-btn.R
\name{rmt_create_btn}
\alias{rmt_create_btn}
\title{Create an \code{RMT3DMS} btn object}
\usage{
rmt_create_btn(
  nlay = 3,
  nrow = 10,
  ncol = 10,
  nper = 1,
  ncomp = 1,
  mcomp = ncomp,
  tunit = "s",
  lunit = "m",
  munit = "g",
  trnop = c(rep(TRUE, 5), rep(FALSE, 5)),
  laycon = rep(1, nlay),
  delr = 100,
  delc = 100,
  htop = 0,
  dz = 10,
  prsity = 0.3,
  icbund = 1,
  sconc = as.list(rep(0, ncomp)),
  cinact = -888,
  thkmin = 0.01,
  ifmtcn = 0,
  ifmtnp = 0,
  ifmtrf = 0,
  ifmtdp = 0,
  savucn = TRUE,
  nprs = -1,
  timprs = NULL,
  obs = NULL,
  nprobs = 1,
  chkmas = TRUE,
  nprmas = 1,
  perlen = rep(1, nper),
  nstp = rep(1, nper),
  tsmult = rep(1, nper),
  ssflag = rep(FALSE, nper),
  tslngh = lapply(seq_len(nper), function(i) rep(1, nstp[i])),
  dt0 = rep(0, nper),
  mxstrn = rep(50, nper),
  ttsmult = rep(1, nper),
  ttsmax = rep(0, nper),
  modflowstylearrays = FALSE,
  drycell = FALSE,
  legacy99storage = FALSE,
  ftlprint = FALSE,
  nowetdryprint = FALSE,
  omitdrycellbudget = FALSE,
  altwtsorb = FALSE
)
}
\arguments{
\item{nlay}{number of layers; defaults to 3}

\item{nrow}{number of rows; defaults to 10}

\item{ncol}{number of columns; defaults to 10}

\item{nper}{number of stress-periods in the flow model; defaults to 1}

\item{ncomp}{total number of chemical species in the simulation; defaults to 1}

\item{mcomp}{total number of mobile species; defaults to \code{ncomp}}

\item{tunit}{character; time unit. Only for identification purposes. Defaults to 's'.}

\item{lunit}{character; length unit. Only for identification purposes. Defaults to 'm'.}

\item{munit}{character: mass unit. Only for identification purposes. Defaults to 'mg'.}

\item{trnop}{logicals denoting which packages are active. Not used by MT3DMS > 4.50 or by MT3D-USGS. Defaults to setting adv, dsp, ssm, rct & gcg active.}

\item{laycon}{vector indicating type of model layer: confined (0) or convertible (not 0; default for all layers).}

\item{delr}{vector of cell widths along rows; defaults to 100 for all columns}

\item{delc}{vector of cell widths along columns; defaults to 100 for all rows}

\item{htop}{matrix with the top elevation of layer 1; defaults to 0 for all cells}

\item{dz}{3D array with the thicknesses of each cell. Defaults to 10 for all cells.}

\item{prsity}{3D array with the (effective or mobile) porosity values for each cell. Defaults to 0.3 for all cells.}

\item{icbund}{3D array with integer values for each cell specifying if solute transport is active (1), inactive (0) or if the cell is a constant-concentration (-1). Defaults to 1 for all cells.}

\item{sconc}{list with \code{ncomp} 3D arrays specifying the starting concentrations in each cell for each species. Defaults to 0 for all cells for all species.}

\item{cinact}{value for indicating an inactive concentration cell (icbund = 0). Defaults to -888.}

\item{thkmin}{decimal fraction of layer thickness below which the cell is considered inactive. Defaults to 0.01. If negative, the absolute value is used.}

\item{ifmtcn}{flag and printing code to determine if calculated concentration should be printed to the standard output text file. Defaults to 0 (no printing).}

\item{ifmtnp}{flag and printing code to determine if number of particles in each cell should be printed to the standard output text file. Defaults to 0 (no printing).}

\item{ifmtrf}{flag and printing code to determine if the model-calculated retardation factor should be printed to the standard output text file. Defaults to 0 (no printing).}

\item{ifmtdp}{flag and printing code to determine if the model-calculated distance-weighted dispersion coefficient should be printed to the standard output text file. Defaults to 0 (no printing).}

\item{savucn}{logical determining if concentration solution should be saved in default unformatted (binary) file named MT3Dnnn.UCN where nnn is the species index number. Defaults to TRUE.}

\item{nprs}{frequency of output. Ignored if \code{timprs} is supplied. If positive, \code{timpr} must be supplied. If negative, output is printed or saved whenever the number of transport steps is an even multiple of \code{nprs}. If zero, only print or save at the end of the simulation. Defaults to -1.}

\item{timprs}{vector with user-defined printing/saving times. Defaults to NULL.}

\item{obs}{optional data.frame with k, i and j columns indicating the cells for which species' concentrations will be saved in MT3Dnnn.OBS files (nnn indicating species index). Defaults to NULL.}

\item{nprobs}{integer indicating how frequently concentrations at \code{obs} observation points should be saved. Concentrations are saved every \code{nprobs} step. Defaults to 1.}

\item{chkmas}{logical indicating whether a one-line summary of the mass balance should be saved to a default MT3Dnnn.MAS file where nnn is the species index. Defaults to TRUE.}

\item{nprmas}{frequency of saving the mass budget information to MT3Dnnn.MAS. Information is saved every \code{nprmas} step. Defaults to 1.}

\item{perlen}{vector of flow model stress-period lengths. Defaults to 1 for every stress-period.}

\item{nstp}{vector of flow model stress-period time steps. Defaults to 1 for every stress-period.}

\item{tsmult}{vector of successive flow model time step length multipliers. Defaults to 1 for every stress-period.}

\item{ssflag}{logical vector of length \code{nper} indicating if the steady-state transport option should be used. Defaults to FALSE for every stress-period. MT3D-USGS only.}

\item{tslngh}{list of length \code{nper} with time step length vectors. Defaults to 1 for every time step length. Only used when \code{tsmult > 0} for the current stress-period.}

\item{dt0}{vector of length \code{nper} of initial transport step sizes within each flow time step. If zero (default) and the GCG solver is active, a model-calculated value based on the Courant number in the Advection package is used.}

\item{mxstrn}{vector of length \code{nper} with the maximum number of transport steps allowed within one flow time step. Defaults to 50 for every stress-period.}

\item{ttsmult}{vector of length \code{nper} with transport step size multipliers within a single flow time step. Defaults to 1 for every stress-period.}

\item{ttsmax}{vector of length \code{nper}with maximum allowed transport step size when \code{ttsmult > 1}. Defaults to 0 for every stress-period (no limit).}

\item{modflowstylearrays}{logical; should MODFLOW-style arrays and array headers be used. Defaults to FALSE. MT3D-USGS only.}

\item{drycell}{logical; should mass transfer through dry cells be enabled when dry cells remain active in the flow simulation (e.g. with MODFLOW-NWT). Only available when mixelm = 0 or -1 in the Advection package. Defaults to FALSE. MT3D-USGS only.}

\item{legacy99storage}{logical; should MT3DMS flow storage computations be used. Defaults to FALSE. MT3D-USGS only.}

\item{ftlprint}{logical; should the contents of the flow-transport link file be printed to the standard output text file. Defaults to FALSE. MT3D-USGS only.}

\item{nowetdryprint}{logical; should printing of messages related to drying and re-wetting of model cells to the standard output text file be disabled? Defaults to FALSE. MT3D-USGS only.}

\item{omitdrycellbudget}{logical; should the printing of the mass budget of dry cells be disabled? Defaults to FALSE. MT3D-USGS only.}

\item{altwtsorb}{logical; use an alternative formulation to simulate adsorbed mass. Defaults to FALSE. MT3D-USGS only.}
}
\value{
an object of class \code{btn}
}
\description{
\code{rmt_create_btn} creates an \code{RMT3DMS} btn object.
}
\examples{
rmt_create_btn()
rmt_create_btn(nper = 3, prsity = array(rep(c(0.1, 0.01, 0.3), each = 100), dim = c(10, 10, 3)))

# multi-species; sconc for species 1 is from rnorm, sconc for species 2 = 0.2
sconc1 <- rmt_create_array(rnorm(100, 0.1, sd = 0.01), dim = c(10, 10, 3))
rmt_create_btn(ncomp = 2, prsity = prsity, sconc = list(sconc1, 0.2))
}
\seealso{
\code{\link{rmt_read_btn}}, \code{\link{rmt_write_btn}}
}
