% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-dsp.R
\name{rmt_create_dsp}
\alias{rmt_create_dsp}
\title{Create an \code{RMT3DMS} dsp object}
\usage{
rmt_create_dsp(
  btn,
  al = 10,
  trpt = 0.1,
  trpv = 0.01,
  dmcoef = 0,
  multidiffusion = FALSE,
  nocross = FALSE
)
}
\arguments{
\item{btn}{\code{RMT3DMS} btn object}

\item{al}{3D array with longitudinal dispersivities. Defaults to 10 for every cell.}

\item{trpt}{1D vector with ratios of horizontal transverse to longitudinal dispersivities (TH/TL) for every layer. Defaults to 0.1 for every layer.}

\item{trpv}{1D vector with ratios of vertical transverse to longitudinal dispersivities (TV/TL) for every layer. Defaults to 0.01 for every layer.}

\item{dmcoef}{either a 1D (nlay) (\code{multidiffusion = FALSE}) or a list of length \code{btn$mcomp} with 3D (\code{multidiffusion = TRUE}) array with the effective molecular diffusion coefficients. Defaults to 0 for all layers or cells (i.e. diffusion is neglected).}

\item{multidiffusion}{logical; if TRUE, \code{dmcoef} can be specified for each species, for each cell. Defaults to FALSE.}

\item{nocross}{logical; should cross-dispersion be disabled? Defaults to FALSE. MT3D-USGS only.}
}
\value{
an object of class \code{dsp}
}
\description{
\code{rmt_create_dsp} creates an \code{RMT3DMS} dsp object.
}
\examples{
btn <- rmt_create_btn(ncomp = 3)
rmt_create_dsp(btn)
rmt_create_dsp(btn, dmcoef = list(0, 0.1, 0.2), multidiffusion = TRUE)
}
\seealso{
\code{\link{rmt_read_dsp}}, \code{\link{rmt_write_dsp}}
}
