% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-gcg.R
\name{rmt_create_gcg}
\alias{rmt_create_gcg}
\title{Create an \code{RMT3DMS} gcg object}
\usage{
rmt_create_gcg(
  mxiter = 1,
  iter1 = 50,
  isolve = 3,
  ncrs = 0,
  accl = 1,
  cclose = 1e-06,
  iprgcg = 0
)
}
\arguments{
\item{mxiter}{maximum number of outer iterations. See details. Defaults to 1.}

\item{iter1}{maximum number of inner iterations. Defaults to 50.}

\item{isolve}{type of preconditioner: 1 = Jacobi, 2 = SSOR, 3 = Modified Incomplete Cholesky (MIC; default).}

\item{ncrs}{integer controlling treatment of dispersion tensor cross-terms. If 0 (default), lump all dispersion cross-terms to RHS. If 1, includes full dispersion tensor.}

\item{accl}{relaxtion factor for the SSOR preconditioner. Defaults to 1.}

\item{cclose}{Convergence criteration in relative concentration. Defaults to 1e-6.}

\item{iprgcg}{interval for printing maximum concentration. changes of each iteration. Defaults to 0 (only print at end of stress-period).}
}
\value{
an object of class \code{gcg}
}
\description{
\code{rmt_create_gcg} creates an \code{RMT3DMS} dsp object.
}
\details{
\code{mxiter} should be higher than 1 when nonlinear sorption isotherm is included or when \code{drycell = TRUE} is used in \code{btn}.
}
\examples{
rmt_create_gcg()
rmt_create_gcg(mxiter = 20, ncrs = 1, cclose = 1e-5, iprgcg = 1)
}
\seealso{
\code{\link{rmt_read_gcg}}, \code{\link{rmt_write_gcg}}
}
