% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmt_create_list}
\alias{rmt_create_list}
\title{Create a rmt_list object}
\usage{
rmt_create_list(df, kper)
}
\arguments{
\item{df}{a data.frame like object with at least columns \code{i, j, k, itype} and one or more \code{css} columns specifying the concentrations}

\item{kper}{integer value(s) specifying during which stress-periods this list is active}
}
\value{
a \code{rmt_list} object
}
\description{
Create a rmt_list object
}
\examples{
df <- data.frame(i = 5, j = 4:5, k = 3, css1 = 6.45, css2 = 47, itype = 2)
rmt_create_list(df, kper = c(1, 3))
}
