% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-nam.R
\name{rmt_create_nam}
\alias{rmt_create_nam}
\title{Create an \code{RMT3DMS} nam object}
\usage{
rmt_create_nam(
  ...,
  ftl = {     cat("Please select corresponding ftl file ...\\n")     file.choose() },
  ftl_free = NULL,
  basename = NULL
)
}
\arguments{
\item{...}{RMT3DMS objects to be included in the nam file}

\item{ftl}{path to the flow-transport link file; typically '.ftl'}

\item{ftl_free}{logical; is the flow-transport link file written in free (formatted) format (TRUE) or binary (unformatted) (FALSE)? if NULL (default), it is guessed from reading \code{ftl}}

\item{basename}{character specifying the basename of the files. The default (\code{NULL}) sets input basenames to 'input' and output to 'output'.}
}
\value{
Object of class mt3d_nam
}
\description{
\code{rmt_create_nam} creates an \code{RMT3DMS} nam object.
}
\details{
If a \code{RMT3DMS nam} object is present, it is recreated. 
 It is advised to place the ftl file in the same directory as the transport model input files.
 The path the flow-transport link file is normalized in \code{rmt_write_nam}.
}
\examples{
btn <- rmt_create_btn()
adv <- rmt_create_adv()
gcg <- rmt_create_gcg()
rmt_create_nam(btn, adv, gcg, ftl = 'output.ftl', ftl_free = TRUE)

}
\seealso{
\code{\link{rmt_read_nam}}, \code{\link{rmt_write_nam}}
}
