% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-ssm.R
\name{rmt_create_ssm}
\alias{rmt_create_ssm}
\title{Create an \code{RMT3DMS} ssm object}
\usage{
rmt_create_ssm(
  ...,
  btn,
  crch = NULL,
  cevt = NULL,
  mxss = prod(btn$nrow, btn$ncol, btn$nlay),
  issgout = NULL
)
}
\arguments{
\item{...}{one or more \code{rmt_list} objects or a single list containing them. These represent point sources/sinks.}

\item{btn}{\code{RMT3DMS} btn object}

\item{crch}{either a single value (used for all species in a multi-species simulation), a single \code{rmt_2d_array} or a list of \code{rmt_2d_arrays}. For multi-species simulations, a list of single values (one for each species) can also be entered. Represents the concentrations of the recharge flux from the flow model.}

\item{cevt}{either a single value (used for all species in a multi-species simulation), a single \code{rmt_2d_array} or a list of \code{rmt_2d_arrays}. For multi-species simulations, a list of single values (one for each species) can also be entered. Represents the concentrations of the evapotranspiration flux from the flow model.}

\item{mxss}{maximum number of all point sources and sinks in the flow model. Defaults to the number of cells. See details.}

\item{issgout}{unit number for an optional output file to save calculated flux-averaged concentrations at multi-node wells. Defaults to NULL. MT3D-USGS only.}
}
\value{
an object of class \code{ssm}
}
\description{
\code{rmt_create_ssm} creates an \code{RMT3DMS} ssm object.
}
\details{
The default value for \code{mxss} is likely to be much larger than the actual number of point sources/sinks. In order to preserve more computer memory, this value can be set to a more appropriate number. 
         All \code{rmt_2d_arrays} should have their kper and solute arguments set through \code{\link{rmt_create_array}}. There can only be one active \code{crch & cevt} \code{rmt_2d_array} per stress-period per species.
         If a negative concentration value/array for \code{crch} or \code{cevt} is specified for the first stress-period, concentrations of positive fluxes are set to zero and concentrations of a negative fluxes to the aquifer concentration for all species. If no other values/arrays are specified, this is repeated for all following stress-periods.
         If a flux is active in the flow simulation, values for the corresponding concentration fluxes must be supplied. E.g. if evt is used in the flow simulation, values for cevt must be supplied even if no concentrations are associated with this flux. In the latter case, supply negative concentration values.
}
\examples{
btn <- rmt_create_btn()
wel <- rmt_create_list(data.frame(i = 5, j = 4:5, k = 3, css1 = 6.45, css2 = 47, itype = 2), kper = 1)
rmt_create_ssm(wel, crch = 0, btn = btn)

btn <- rmt_create_btn(nper = 3, ncomp = 2)
wel_df <- data.frame(i = 5, j = 4:5, k = 3, css1 = 6.45, css2 = 47, itype = 2)
wel <- rmt_create_list(wel_df, kper = c(1,3))
wel2_df <- wel_df
wel2_df$css1 <- 0
wel2 <- rmt_create_list(wel2_df, kper = 2)
riv <- rmt_create_list(data.frame(i = 1, j = 1:10, k = 1, css1 = 12, css2 = 0, itype = 4), kper = c(1,2))
evt <- list(rmt_create_array(0, dim = c(btn$nrow, btn$ncol), solute = c(1,2), kper = c(1,3)),
            rmt_create_array(12, dim=c(btn$nrow, btn$ncol), solute = 1, kper = 2),
            rmt_create_array(0, dim=c(btn$nrow, btn$ncol), solute = 2, kper = 2))
rmt_create_ssm(wel, wel2, riv, crch = list(0, 2), cevt = evt, btn = btn)

rmt_create_ssm(cevt = -1, btn = btn)

}
\seealso{
\code{\link{rmt_read_ssm}}, \code{\link{rmt_write_ssm}}
}
