% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{rmt_execute}
\alias{rmt_execute}
\alias{rmt_execute.character}
\alias{rmt_execute.mt3dms}
\title{Execute a MT3DMS or MT3D-USGS model}
\usage{
rmt_execute(...)

\method{rmt_execute}{character}(
  path,
  code = "mt3d-usgs",
  ui = NULL,
  convergence = "Program completed",
  bits = "single",
  form = "b"
)

\method{rmt_execute}{mt3dms}(
  mt3dms,
  code = "mt3d-usgs",
  ui = NULL,
  convergence = "Program completed",
  bits = "single",
  form = "b"
)
}
\arguments{
\item{path}{Path to the NAM file. Typically with extension `.nam`.}

\item{code}{Name of the MT3D variant to use, or path to the executable.}

\item{ui}{If NULL (default), MT3D output is shown in the R console. If
`"none"`, the output is suppressed.}

\item{convergence}{Character. The message in the terminal output used
to check for convergence.}

\item{bits}{Character. Can be \code{"single"} (ie. 32 bits) or \code{"double"} (i.e. 64 bits). Specifies if 32 or 64 bit executable are used for the MT3D-USGS executable.}

\item{form}{Character. Can be \code{'s'} or \code{'b'} (default). Specifies which MT3DMS executable is used, which depends on the value of FORM in the 'filespec.inc' file during compilation. The default 'b' (for binary) should work with FTL files generated by MODFLOW-2000 and newer versions. A value of 's' (for unformatted) should work with FTL files generated by MODFLOW-88 or 96.}

\item{mt3dms}{mt3dms object}
}
\value{
Invisible list with elapsed run time, a logical
  indicating normal termination, and the stdout output, when done for an on
  disk model. Full `mt3dms` object including all results otherwise (in
  memory model).
}
\description{
These functions execute MT3DMS or MT3D-USGS models.
}
\seealso{
[rmt_install()] for external code installation
}
