% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmt_plot.cbud}
\alias{rmt_plot.cbud}
\title{Plot a MT3DMS mass budget}
\usage{
\method{rmt_plot}{cbud}(
  cbud,
  btn,
  icomp = 1,
  what = "cumulative",
  fluxes = "all",
  net = FALSE,
  type = "area",
  final = FALSE
)
}
\arguments{
\item{cbud}{\code{RMT3DMS} cbud object as obtained by \code{rmt_read_bud}}

\item{btn}{\code{RMT3DMS} btn object}

\item{icomp}{integer selecting for which species to plot the budget; defaults to 1}

\item{what}{character; what to plot, "cumulative", "total", "difference" or "discrepancy". Defaults to "cumulative".}

\item{fluxes}{character; either "all" or a character vector with the flux components to plot. Only used when \code{what} is "cumulative"}

\item{net}{logical; if TRUE, it sums the inflows and outflows of the flux component to plot the net fluxes. If FALSE, it will plot both the inflows and outflows. Only used when \code{what} is "cumulative" or "total".}

\item{type}{character; plot type. Either "bar" or "area".}

\item{final}{logical; should only the final mass budget be plotted? Defaults to FALSE}
}
\value{
ggplot2 object
}
\description{
\code{rmt_plot.cbud} plots a MT3DMS mass budget
}
