% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmt_plot.cobs}
\alias{rmt_plot.cobs}
\title{Plot a MT3DMS concentrations observation time series}
\usage{
\method{rmt_plot}{cobs}(cobs, i = NULL, j = NULL, k = NULL)
}
\arguments{
\item{cobs}{\code{RMT3DMS} cobs objected as returned from \code{rmt_read_obs}}

\item{i}{row number to subset observations. Defaults to NULL.}

\item{j}{column number subset observations. Defaults to NULL.}

\item{k}{layer number subset observations. Defaults to NULL.}
}
\value{
ggplot2 object
}
\description{
Plot a MT3DMS concentrations observation time series
}
\details{
The observations are named using their k-i-j indices.
}
