% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmt_plot.mas}
\alias{rmt_plot.mas}
\title{Plot a MT3DMS mass summary file time series}
\usage{
\method{rmt_plot}{mas}(mas, variable = "all", final = FALSE)
}
\arguments{
\item{mas}{\code{RMT3DMS} mas object as returned from \code{rmt_read_mas}}

\item{variable}{character specifying which variable to plot. Defaults to 'all'. Allowed values are the colnames of \code{mas}.}

\item{final}{logical, should only the final time step be plotted. Default to FALSE.}
}
\value{
ggplot2 object
}
\description{
Plot a MT3DMS mass summary file time series
}
\details{
plots a time series line plot of the variable. If \code{variable = 'all'}, \code{ggplot2::facet_wrap} is used to plot all variable with free y scales. 
 If \code{final = TRUE}, a bar plot is returned using \code{ggplot2::geom_col} of the final time step.
}
