% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmt_plot.mfx}
\alias{rmt_plot.mfx}
\title{Plot a MT3DMS mass flux file}
\usage{
\method{rmt_plot}{mfx}(mfx, type = "scatter", bins = NULL)
}
\arguments{
\item{mfx}{\code{RMT3DMS} mfx object}

\item{type}{plot type: 'scatter', 'residual', 'time' or 'histogram'}

\item{bins}{number of bins to use in the histogram plot; defaults to the Freedman-Diaconis rule}
}
\description{
Plot a MT3DMS mass flux file
}
