% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmt_plot.ocn}
\alias{rmt_plot.ocn}
\title{Plot a MT3DMS observed concentration file}
\usage{
\method{rmt_plot}{ocn}(ocn, type = "scatter", cinact = -888, bins = NULL)
}
\arguments{
\item{ocn}{\code{RMT3DMS} ocn object}

\item{type}{plot type: 'scatter', 'residual', 'time' or 'histogram'}

\item{cinact}{value used to flag dry cells; defaults to -888}

\item{bins}{number of bins to use in the histogram plot; defaults to the Freedman-Diaconis rule}
}
\description{
Plot a MT3DMS observed concentration file
}
