% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmt_plot.rmt_2d_array}
\alias{rmt_plot.rmt_2d_array}
\title{Plot a MT3DMS 2D array}
\usage{
\method{rmt_plot}{rmt_2d_array}(
  rmt_2d_array,
  btn,
  mask = {     warning("Using first icbund layer as mask.", call. = FALSE)    
    btn$icbund[, , 1] },
  ...
)
}
\arguments{
\item{rmt_2d_array}{an object of class rmt_2d_array}

\item{btn}{basic transport file object}

\item{...}{arguments passed to RMODFLOW::rmf_plot.rmf_2d_array}
}
\value{
ggplot2 object or layer; if plot3D is TRUE, nothing is returned and the plot is made directly
}
\description{
\code{rmt_plot.rmt3dms_2d_array} plots a MODFLOW 2D array.
}
