% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmt_plot.rmt_4d_array}
\alias{rmt_plot.rmt_4d_array}
\title{Plot a 2D section through a MT3DMS 4D array}
\usage{
\method{rmt_plot}{rmt_4d_array}(
  rmt_4d_array,
  btn,
  i = NULL,
  j = NULL,
  k = NULL,
  l = NULL,
  mask = btn$icbund,
  ...
)
}
\arguments{
\item{rmt_4d_array}{an object of class rmt_4d_array}

\item{btn}{basic transport file object}

\item{i}{row number to plot}

\item{j}{column number to plot}

\item{k}{layer number to plot}

\item{l}{transport time step number to plot (index for 4th dimension)}

\item{mask}{a 3D array with 0 or F indicating inactive cells; defaults to btn$icbund}

\item{...}{arguments provided to RMODFLOW::rmf_plot.rmf_3d_array}
}
\value{
ggplot2 object or layer; if plot3D is TRUE, nothing is returned and the plot is made directly
}
\description{
\code{rmt_plot.rmt_4d_array} plots a 2D section through a MT3DMS 4D array.
}
