% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmt_plot.rmt_list}
\alias{rmt_plot.rmt_list}
\title{Plot a RMT3DMS list object}
\usage{
\method{rmt_plot}{rmt_list}(
  obj,
  btn,
  mask = btn$icbund,
  variable = "id",
  i = NULL,
  j = NULL,
  k = NULL,
  ...
)
}
\arguments{
\item{obj}{a \code{RMT3DMS} object of class \code{rmt_list}}

\item{btn}{a \code{RMT3DMS} btn object}

\item{mask}{a 3D array with 0 or F indicating inactive cells; defaults to btn$icbund}

\item{variable}{single character or numeric indicating which column in the \code{rmt_list} object to plot. Defaults to 'id', which plots the locations of the cells.}

\item{i}{row number to plot}

\item{j}{column number to plot}

\item{k}{layer number to plot}

\item{...}{additional arguments passed to either \code{\link{RMODFLOW::rmf_plot.rmf_list}}}
}
\value{
ggplot2 object or layer
}
\description{
Plot a RMT3DMS list object
}
