% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmt_plot.tob}
\alias{rmt_plot.tob}
\title{Plot a RMT3DMS tob object}
\usage{
\method{rmt_plot}{tob}(
  tob,
  btn,
  i = NULL,
  j = NULL,
  k = NULL,
  what = "concentrations",
  prj = RMODFLOW::rmf_get_prj(btn),
  ...
)
}
\arguments{
\item{tob}{\code{RMT3DMS} tob object}

\item{btn}{\code{RMT3DMS} btn object}

\item{i}{row number to plot}

\item{j}{column number to plot}

\item{k}{layer number to plot}

\item{what}{character, either 'concentrations' (default) or 'fluxes'. Denotes which observations to plot.}

\item{prj}{projection file object}

\item{...}{additional arguments passed to \code{\link{rmt_plot.rmt_list}}}
}
\value{
ggplot2 object or layer
}
\description{
Plot a RMT3DMS tob object
}
\details{
specifying all of the \code{i, j & k} arguments will plot a time series at that cell location. The observations are grouped by name.
}
