% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt3dms.R
\name{rmt_read}
\alias{rmt_read}
\title{Read a MT3DMS model}
\usage{
rmt_read(
  file = {     cat("Please select nam file ...\\n")     file.choose() },
  output = FALSE,
  precision = "single",
  verbose = TRUE
)
}
\arguments{
\item{file}{NAME file; typically '*.mt_nam'}

\item{output}{logical; should output also be read. Defaults to FALSE.}

\item{precision}{either \code{'single'} or \code{'double'}. Specifies the precision of binary files. Defaults to \code{'single'}.}

\item{verbose}{logical; should information on reading files be printed to the console ? Defaults to TRUE.}
}
\value{
a \code{mt3dms} object which is a list containing all MT3DMS packages and optionally, model output
}
\description{
\code{rmt_read} reads in a MT3DMS/MT3D-USGS model and returns it as a \code{mt3dms} object
}
\note{
If the MT3DMS model does not contain a NAME file, use the \code{rmt_read_*} functions to read in the packages individually.
}
\seealso{
\code{\link{rmt_create}}, \code{\link{rmt_write}}
}
