% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-adv.R
\name{rmt_read_adv}
\alias{rmt_read_adv}
\title{Read an MT3DMS advection package file}
\usage{
rmt_read_adv(
  file = {     cat("Please select adv file ...\\n")     file.choose() }
)
}
\arguments{
\item{file}{filename; typically '*.adv'}
}
\value{
object of class adv
}
\description{
\code{rmt_read_adv} reads in an MT3DMS advection package file and returns it as an \code{\link{RMT3DMS}} adv object.
}
\seealso{
\code{\link{rmt_create_adv}}, \code{\link{rmt_write_adv}}
}
