% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-btn.R
\name{rmt_read_btn}
\alias{rmt_read_btn}
\title{Read an MT3DMS basic transport package file}
\usage{
rmt_read_btn(
  file = {     cat("Please select btn file ...\\n")     file.choose() },
  ...
)
}
\arguments{
\item{file}{filename; typically '*.btn'}

\item{...}{optional arguments passed to \code{\link{rmti_parse_array}}}
}
\value{
object of class btn
}
\description{
\code{rmt_read_btn} reads in an MT3DMS basic transport package file and returns it as an \link{RMT3DMS} btn object.
}
