% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{rmt_read_bud}
\alias{rmt_read_bud}
\title{Reads the mass budget from a MT3DMS listing file}
\usage{
rmt_read_bud(
  file = {     cat("Please select MT3DMS listing file ...\\n")     file.choose() }
)
}
\arguments{
\item{file}{path to the listing file; typically '*.m3d'}
}
\value{
an object of class cbud which is a data.frame containing the cumulative mass budgets for all species
}
\description{
\code{rmt_read_bud} reads a mass budget from a MT3DMS listing file and returns it as a data frame
}
