% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-gcg.R
\name{rmt_read_gcg}
\alias{rmt_read_gcg}
\title{Read an MT3DMS generalized conjugate gradient solver package file}
\usage{
rmt_read_gcg(
  file = {     cat("Please select gcg file ...\\n")     file.choose() }
)
}
\arguments{
\item{file}{filename; typically '*.gcg'}
}
\value{
object of class gcg
}
\description{
\code{rmt_read_gcg} reads in an MT3DMS generalized conjugate gradient solver package file and returns it as an \code{\link{RMT3DMS}} gcg object.
}
