% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{rmt_read_mas}
\alias{rmt_read_mas}
\title{Read an MT3DMS mass balance summary file}
\usage{
rmt_read_mas(
  file = {     cat("Please select mas file ...\\n")     file.choose() }
)
}
\arguments{
\item{file}{filename; typically '*.mas'}
}
\value{
object of class mas
}
\description{
\code{rmt_read_mas} reads in an MT3DMS mass balance summary file and returns it as an \code{RMT3DMS} mas object.
}
