% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{rmt_read_mfx}
\alias{rmt_read_mfx}
\title{Read a MT3DMS mass flux file}
\usage{
rmt_read_mfx(
  file = {     cat("Please select mfx file ...\\n")     file.choose() }
)
}
\arguments{
\item{file}{filename; typically '*.mfx'}
}
\value{
object of class mfx and data.frame
}
\description{
\code{rmt_read_mfx} reads in an MT3DMS mass flux file and returns it as an \code{\link{RMT3DMS}} mfx object.
}
\details{
Writing and type of MFX output is set in the TOB file.
}
\seealso{
\code{\link{rmt_read_mfx}}, \code{\link{rmt_read_pst}} and \code{\link{rmt_create_tob}}
}
