% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-nam.R
\name{rmt_read_nam}
\alias{rmt_read_nam}
\title{Read a MT3DMS name file}
\usage{
rmt_read_nam(
  file = {     cat("Please select nam file ...\\n")     file.choose() }
)
}
\arguments{
\item{file}{filename; typically '*.nam'}
}
\value{
object of class mt3d_nam
}
\description{
\code{rmt_read_nam} reads in a MT3DMS name file and returns it as an \code{\link{RMT3DMS}} nam object.
}
\seealso{
\code{\link{rmt_read_nam}}, \code{\link{rmt_write_nam}}
}
