% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{rmt_read_obs}
\alias{rmt_read_obs}
\title{Read an MT3DMS Concentrations Observation File}
\usage{
rmt_read_obs(
  file = {     cat("Please select obs file ...\\n")     file.choose() },
  btn = {     cat("Please select corresponding btn file ...\\n")    
    rmt_read_btn(file.choose()) },
  solute = NULL
)
}
\arguments{
\item{file}{filename; typically '*.obs'}

\item{btn}{\code{RMT3DMS} btn object}

\item{solute}{optional integer used to set the species index in. By default, the code tries to guess this from the filename (e.g. 'MT3D001.OBS')}
}
\value{
object of class cobs
}
\description{
\code{rmt_read_obs} reads in an MT3DMS concentration observation file and returns it as an \code{\link{RMT3DMS}} cobs object.
}
