% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{rmt_read_ocn}
\alias{rmt_read_ocn}
\title{Read a MT3DMS observed concentration file}
\usage{
rmt_read_ocn(
  file = {     cat("Please select ocn file ...\\n")     file.choose() }
)
}
\arguments{
\item{file}{filename; typically '*.ocn'}
}
\value{
object of class ocn and data.frame
}
\description{
\code{rmt_read_ocn} reads in an MT3DMS observed concentration file and returns it as an \code{\link{RMT3DMS}} ocn object.
}
\details{
Writing and type of OCN output is set in the TOB file.
}
\seealso{
\code{\link{rmt_read_mfx}}, \code{\link{rmt_read_pst}} and \code{\link{rmt_create_tob}}
}
