% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-tob.R
\name{rmt_read_tob}
\alias{rmt_read_tob}
\title{Read an MT3DMS transport observation package file}
\usage{
rmt_read_tob(
  file = {     cat("Please select tob file ...\\n")     file.choose() }
)
}
\arguments{
\item{file}{filename; typically '*.tob'}
}
\value{
object of class tob
}
\description{
\code{rmt_read_tob} reads in an MT3DMS transport observation package file and returns it as an \code{\link{RMT3DMS}} tob object.
}
\seealso{
\code{\link{rmt_create_tob}}, \code{\link{rmt_write_tob}}
}
