% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{rmt_read_ucn}
\alias{rmt_read_ucn}
\title{Read an MT3DMS unformatted concentration file}
\usage{
rmt_read_ucn(
  file = {     cat("Please select ucn file ...\\n")     file.choose() },
  btn = {     cat("Please select corresponding btn file ...\\n")    
    rmt_read_btn(file.choose()) },
  mask = btn$icbund,
  cinact = btn$cinact,
  solute = NULL,
  precision = "single"
)
}
\arguments{
\item{file}{filename; typically '*.ucn'}

\item{btn}{\code{RMT3DMS} btn object}

\item{mask}{3d array which can be coerced to logical indicating active (TRUE) or inactive (FALSE) cells. Concentration of inactive cells is set to NA. Defaults to the icbund array in btn.}

\item{cinact}{concentration value(s) indicating inactive cells. These cells are set to NA. Defaults to cinact in btn.}

\item{solute}{optional integer used to set the species index in. By default, the code tries to guess this from the filename (e.g. 'MT3D001.UCN')}

\item{precision}{either \code{'single'} (default) or \code{'double'}. Specifies the precision of the binary file.}
}
\value{
object of class ucn and rmt_4d_array
}
\description{
\code{rmt_read_ucn} reads in an MT3DMS unformatted concentration file and returns it as an \code{\link{RMT3DMS}} ucn object.
}
