% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{rmt_run_mt3dms}
\alias{rmt_run_mt3dms}
\title{Run an MT3DMS model}
\usage{
rmt_run_mt3dms(file, mt3dms_executable = "mt3dms5b")
}
\arguments{
\item{file}{path to name file; typically '*.nam'}

\item{mt3dms_executable}{name of the MT3DMS executable to use}
}
\description{
\code{run_mt3dms} runs an MT3DMS model.
}
