% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{rmt_run_mt3dms_ss}
\alias{rmt_run_mt3dms_ss}
\title{Run an MT3DMS model}
\usage{
rmt_run_mt3dms_ss(
  file,
  threshold,
  report = TRUE,
  report_plot = TRUE,
  maxit = 100,
  ss = NULL,
  ...
)
}
\arguments{
\item{file}{path to name file; typically '*.nam'}

\item{threshold}{threshold giving the maximum concentration difference for two consecutive simulations}

\item{report}{logical, should information be printed after each iteration}

\item{report_plot}{logical, should information be plotted after each iteration}

\item{maxit}{maximum number of iterations}

\item{ss}{ss object, to continue a simulation from a previous call to \code{run_mt3dms_ss}}

\item{...}{parameters passed to \code{run_mt3dms}}
}
\description{
\code{run_mt3dms_ss} keeps on running an MT3DMS model until convergence is reached
}
