% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-adv.R
\name{rmt_write_adv}
\alias{rmt_write_adv}
\title{Write an MT3DMS advection package file}
\usage{
rmt_write_adv(
  adv,
  file = {    
    cat("Please select adv file to overwrite or provide new filename ...\\n")    
    file.choose() }
)
}
\arguments{
\item{adv}{an \code{RMT3DMS} adv object}

\item{file}{filename to write to; typically '*.adv'}
}
\value{
\code{NULL}
}
\description{
Write an MT3DMS advection package file
}
\seealso{
\code{\link{rmt_read_adv}}, \code{\link{rmt_create_adv}}
}
