% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-btn.R
\name{rmt_write_btn}
\alias{rmt_write_btn}
\title{Write a MT3DMS basic transport package file}
\usage{
rmt_write_btn(
  btn,
  file = {    
    cat("Please select btn file to overwrite or provide new filename ...\\n")    
    file.choose() },
  iprn = -1,
  ...
)
}
\arguments{
\item{btn}{an \code{RMT3DMS} btn object}

\item{file}{filename to write to; typically '*.btn'}

\item{iprn}{format code for printing arrays in the listing file; defaults to -1 (no printing)}

\item{...}{arguments passed to \code{rmti_write_array}.}
}
\value{
\code{NULL}
}
\description{
Write a MT3DMS basic transport package file
}
\seealso{
\code{\link{rmt_read_btn}}, \code{\link{rmt_create_btn}}
}
