% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-dsp.R
\name{rmt_write_dsp}
\alias{rmt_write_dsp}
\title{Write an MT3DMS dispersion package file}
\usage{
rmt_write_dsp(
  dsp,
  file = {    
    cat("Please select dsp file to overwrite or provide new filename ...\\n")    
    file.choose() },
  btn = {     cat("Please select corresponding btn file ...\\n")    
    rmt_read_btn(file.choose()) },
  iprn = -1,
  ...
)
}
\arguments{
\item{dsp}{an \code{RMT3DMS} dsp object}

\item{file}{filename to write to; typically '*.dsp'}

\item{btn}{an \code{RMT3DMS} btn object}

\item{iprn}{format code for printing arrays in the listing file; defaults to -1 (no printing)}

\item{...}{arguments passed to \code{rmti_write_array}.}
}
\value{
\code{NULL}
}
\description{
Write an MT3DMS dispersion package file
}
\seealso{
\code{\link{rmt_read_dsp}}, \code{\link{rmt_create_dsp}}
}
