% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-gcg.R
\name{rmt_write_gcg}
\alias{rmt_write_gcg}
\title{Write an MT3DMS generalized conjugate gradient solver package file}
\usage{
rmt_write_gcg(
  gcg,
  file = {    
    cat("Please select gcg file to overwrite or provide new filename ...\\n")    
    file.choose() }
)
}
\arguments{
\item{gcg}{an \code{RMT3DMS} gcg object}

\item{file}{filename to write to; typically '*.gcg'}
}
\value{
\code{NULL}
}
\description{
Write an MT3DMS generalized conjugate gradient solver package file
}
