% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-rct.R
\name{rmt_write_rct}
\alias{rmt_write_rct}
\title{Write an MT3DMS Chemical Reaction Package file}
\usage{
rmt_write_rct(
  rct,
  file = {    
    cat("Please select rct file to overwrite or provide new filename ...\\n")    
    file.choose() },
  btn = {     cat("Please select corresponding btn file ...\\n")    
    rmt_read_btn(file.choose()) },
  iprn = -1,
  ...
)
}
\arguments{
\item{rct}{an \code{RMT3DMS} rct object}

\item{file}{filename to write to; typically '*.rct'}

\item{btn}{an \code{RMT3DMS} btn object}

\item{iprn}{format code for printing arrays in the listing file; defaults to -1 (no printing)}

\item{...}{arguments passed to \code{rmti_write_array}.}
}
\value{
\code{NULL}
}
\description{
Write an MT3DMS Chemical Reaction Package file
}
