% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-ssm.R
\name{rmt_write_ssm}
\alias{rmt_write_ssm}
\title{Write an MT3DMS Sink-Source Mixing Package file}
\usage{
rmt_write_ssm(
  ssm,
  file = {    
    cat("Please select ssm file to overwrite or provide new filename ...\\n")    
    file.choose() },
  btn = {     cat("Please select corresponding btn file ...\\n")    
    rmt_read_btn(file.choose()) },
  iprn = -1,
  ...
)
}
\arguments{
\item{ssm}{an \code{RMT3DMS} ssm object}

\item{file}{filename to write to; typically '*.ssm'}

\item{btn}{an \code{RMT3DMS} btn object}

\item{iprn}{format code for printing arrays in the listing file; defaults to -1 (no printing)}

\item{additional}{arguments passed to \code{\link{rmti_write_array}}}
}
\value{
\code{NULL}
}
\description{
\code{rmt_write_ssm} writes an MT3DMS sink & source mixing package file from a \link{RMT3DMS} ssm object.
}
\examples{
btn <- rmt_create_btn()
wel <- rmt_create_list(data.frame(i = 5, j = 4:5, k = 3, css1 = 6.45, css2 = 47, itype = 2), kper = 1)
ssm <- rmt_create_ssm(wel, crch = 0, btn = btn)
rmt_write_ssm(ssm, file = file.path(tempdir(), 'input.ssm'), btn = btn)
}
\seealso{
\code{\link{rmt_create_ssm}}, \code{\link{rmt_read_ssm}}
}
