% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmti_check_ftl_binary}
\alias{rmti_check_ftl_binary}
\title{Check if the FTL file is binary}
\usage{
rmti_check_ftl_binary(file)
}
\arguments{
\item{file}{pathname to the flow-transport link file, typically '*.ftl'}
}
\value{
logical indicating if the FTL file is binary or keyword 'empty' if file is empty.
}
\description{
Check if the FTL file is binary
}
\keyword{internal}
