% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmti_create_bc_list}
\alias{rmti_create_bc_list}
\title{Function to create a rmt_list object from RMODFLOW discete boundary condition object objects}
\usage{
rmti_create_bc_list(obj, conc, itype, kper)
}
\arguments{
\item{obj}{RMODFLOW discete boundary condition object. Allowed objects are chd, ghb, riv, drn, wel}

\item{conc}{vector, data.frame or matrix with concentration values.}

\item{itype}{integer column (or single value) to indicate flux type}

\item{kper}{integer value(s) indicating during which stress-periods this rmf_list is active}
}
\value{
a \code{rmt_list} object
}
\description{
Function to create a rmt_list object from RMODFLOW discete boundary condition object objects
}
\keyword{internal}
