% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{rmti_download_code}
\alias{rmti_download_code}
\title{Download a code}
\usage{
rmti_download_code(code, dir, os, overwrite)
}
\arguments{
\item{code}{Character vector with the codes to install, or `"all"` (default).}

\item{dir}{Installation directory.}

\item{os}{Operating system.}

\item{overwrite}{Logical. Overwrite when the code is already installed? If
`NULL` (default), the user is asked what to do in an interactive session.
An error message is issued otherwise.}
}
\description{
Download a code
}
