% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{rmti_find}
\alias{rmti_find}
\title{Find paths to executables}
\usage{
rmti_find(code = "usgs", bits = "single", form = "b")
}
\arguments{
\item{code}{Name of the MT3D variant to use, or path to the executable.}

\item{bits}{Character. Can be \code{"single"} (ie. 32 bits) or \code{"double"} (i.e. 64 bits). Specifies if 32 or 64 bit executable are used for the MT3D-USGS executable.}

\item{form}{Character. Can be \code{'s'} or \code{'b'} (default). Specifies which MT3DMS executable is used, which depends on the value of FORM in the 'filespec.inc' file during compilation. The default 'b' (for binary) should work with FTL files generated by MODFLOW-2000 and newer versions. A value of 's' (for unformatted) should work with FTL files generated by MODFLOW-88 or 96.}
}
\value{
Path to the executable.
}
\description{
This function tries to locate external code executables.
}
\details{
It first looks for the executable in the current working directory. If not
there, it looks in the bin subfolder of `getOption("RMT3DMS.path")`, where
the software might have been installed by [rmt_install()]. If the executable
cannot be found, a final attempt is made by checking the system path
variable. If it still cannot be located, an error is thrown.
}
