% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmti_ifelse0}
\alias{rmti_ifelse0}
\title{Conditional return}
\usage{
rmti_ifelse0(test, yes, no)
}
\arguments{
\item{test}{an object which can be coerced to logical mode.}

\item{yes}{return value for \code{test==TRUE}}

\item{no}{return value for \code{test==FALSE}}
}
\description{
\code{rmti_ifelse0} returns \code{yes} if \code{test} is \code{TRUE}. If \code{test} is \code{FALSE}, it returns \code{no}.
}
\keyword{internal}
