% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{rmti_install_code}
\alias{rmti_install_code}
\title{Install codes}
\usage{
rmti_install_code(code, overwrite)
}
\arguments{
\item{code}{Character vector with the codes to install, or `"all"` (default).}

\item{overwrite}{Logical. Overwrite when the code is already installed? If
`NULL` (default), the user is asked what to do in an interactive session.
An error message is issued otherwise.}
}
\description{
Install codes
}
