% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmti_itype}
\alias{rmti_itype}
\title{Return a data.frame with itype and names columns}
\usage{
rmti_itype()
}
\value{
a data.frame containing the 'itype' column with the integer code used by MT3D to define flux types and the 'names' column with corresponding flux names.
}
\description{
Return a data.frame with itype and names columns
}
\keyword{internal}
