% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmti_list_packages}
\alias{rmti_list_packages}
\title{List supported MT3DMS/MT3D-USGS packages}
\usage{
rmti_list_packages(type = "usgs")
}
\arguments{
\item{type}{character denoting type of packages to list; possible values are \code{'usgs' (default), 'mt3dms'}}
}
\value{
data.frame with ftype and rmt columns denoting the MT3DMS and \code{RMT3DMS} abbreviations for the requested packages
}
\description{
List supported MT3DMS/MT3D-USGS packages
}
\details{
'usgs' holds all packages; 'mt3dms'
}
\note{
this function should be updated every time a new MT3DMS package is supported in \code{RMT3DMS}
}
\keyword{internal}
