% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmti_parse_array}
\alias{rmti_parse_array}
\title{Get an array specified by a control record from the text lines analyzed in a \code{RMT3DMS} \code{rmt_read_*} function}
\usage{
rmti_parse_array(
  remaining_lines,
  nrow,
  ncol,
  nlay,
  ndim,
  skip_header = FALSE,
  nam = NULL,
  precision = "single",
  file = NULL,
  integer = FALSE,
  ...
)
}
\arguments{
\item{remaining_lines}{lines to read the array from}

\item{nrow}{number of rows in the array}

\item{ncol}{number of columns in the array}

\item{nlay}{number of layers in the array that should be read}

\item{ndim}{dimensions of the array to read; either 1, 2 or 3. Denotes the if the returned array should be 1D, 2D or 3D.}

\item{skip_header}{optional; should the control record be skipped}

\item{nam}{a \code{RMT3DMS} nam object. Required when reading external arrays}

\item{precision}{character: either \code{'single'} (default) or \code{'double'}. Denotes the precision of binary files}

\item{file}{pathname to the input file which is currently being read. Required when reading fixed-format or MODFLOW-style OPEN/CLOSE arrays}

\item{integer}{logical; does the binary array hold integer values. Might not work optimally.}

\item{...}{ignored}
}
\value{
A list containing the array and the remaining text of the MT3DMS input file
}
\description{
Get an array specified by a control record from the text lines analyzed in a \code{RMT3DMS} \code{rmt_read_*} function
}
\keyword{internal}
