% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmti_parse_comments}
\alias{rmti_parse_comments}
\title{Read comments
Internal function used in the read_* functions to read comments}
\usage{
rmti_parse_comments(remaining_lines, id = NULL)
}
\arguments{
\item{id}{optional integers specifying which lines are comments. If NULL (default), lines starting with "#" indicate commented lines}
}
\description{
Read comments
Internal function used in the read_* functions to read comments
}
\details{
removes empty comments and prevents copying of RMT3DMS header comment
}
\keyword{internal}
