% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmti_parse_ftl_header}
\alias{rmti_parse_ftl_header}
\title{Read the package header from a flow-transport link file}
\usage{
rmti_parse_ftl_header(file, binary = NULL)
}
\arguments{
\item{file}{pathname to the flow-transport link file, typically '*.ftl'}

\item{binary}{is the FTL file binary?}
}
\value{
logical vector of length 3, indicating if the rch, evt or uzf packages are active in the flow simulation
}
\description{
Read the package header from a flow-transport link file
}
\details{
This function is used in \code{\link{rmt_read_ssm}} as a replacement for reading data set 1. It is used to determine if
         rch, evt and/or uzf concentration arrays have to be read. All other active flow packages are read from the point source/sink data sets.
         \code{rmti_parse_ftl_header} can only be used with flow-transport link files using MT3DMS or MT3D-USGS headers.
         The UZF package can only be used with MT3D-USGS.
}
\keyword{internal}
