% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmti_parse_variables}
\alias{rmti_parse_variables}
\title{Read MT3DMS single-line variables}
\usage{
rmti_parse_variables(
  remaining_lines,
  n,
  width = 10,
  nlay = NULL,
  character = FALSE,
  format = "fixed",
  ...
)
}
\arguments{
\item{n}{integer; number of variables to be returned. Only used when format is \code{'fixed'}.}

\item{width}{integer; length of a single variable. Only used when format is \code{'fixed'}.}

\item{nlay}{integer; number of layers for which values are to be read. Only use when a 1D(NLAY) variable is read which may be specified on multiple lines.}

\item{character}{logical; should a character vector be returned. Prevents conversion from character names to numeric. Defaults to FALSE. Useful if only characters are present on the line.}

\item{format}{character, either \code{'free'} or \code{'fixed'}. When 'fixed', reads character fields of length 'width' and converts to numeric. Empty fields are set to zero.}

\item{...}{ignored}
}
\value{
vector with values
}
\description{
Read MT3DMS single-line variables
}
\keyword{internal}
