% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmti_write_array}
\alias{rmti_write_array}
\title{Write MT3DMS array
Internal function used in the rmt_write_* functions for writing array datasets}
\usage{
rmti_write_array(
  array,
  file,
  mf_style = FALSE,
  format = "free",
  cnstnt = 1,
  iprn = -1,
  append = TRUE,
  ...
)
}
\arguments{
\item{array}{array to write}

\item{file}{pathname to the file to write the array to}

\item{mf_style}{logical, should MODFLOW-style array headers be used (i.e. INTERNAL, EXTERNAL, OPEN/CLOSE, ...) ? Defaults to FALSE}

\item{format}{either 'free' (iread = 103, i.e. FORTRAN free format) or 'fixed' (iread = 100 using FORTRAN format 10G11.4). In both cases, iread = 0 when the array only contains 1 unique value. Defaults to 'free'.}

\item{cnstnt}{numeric constant to add to the array header which acts as a multiplier for the array values in MODFLOW. Default to 1}

\item{iprn}{iprn code to add to array header. Defaults to -1}

\item{append}{logical; should array be appended to the file. Defaults to TRUE}

\item{...}{passed to \code{\link{RMODFLOW:::rmfi_write_array}} when mf_style is TRUE}
}
\description{
Write MT3DMS array
Internal function used in the rmt_write_* functions for writing array datasets
}
\keyword{internal}
