% Generated by roxygen2 (4.0.2): do not edit by hand
\name{run_mt3dms}
\alias{run_mt3dms}
\title{Run an MT3DMS model}
\usage{
run_mt3dms(nam, dir = getwd(), mt3dmsVersion = "mt3dms5b")
}
\arguments{
\item{file}{Filename; typically *.btn}
}
\value{
Object of class btn
}
\description{
\code{read.btn} reads in an MT3DMS basic transport package file and returns it as an \code{\link{RMT3DMS}} btn object.
}

