% Generated by roxygen2 (4.0.2): do not edit by hand
\name{run_mt3dms_ss}
\alias{run_mt3dms_ss}
\title{Run an MT3DMS model}
\usage{
run_mt3dms_ss(file, threshold, report = TRUE, report_plot = TRUE, ...)
}
\arguments{
\item{file}{Path to name file; typically "*.nam"}
}
\description{
\code{run_mt3dms_ss} keeps on running an MT3DMS model until convergence is reached
}

