% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_adv.R
\name{write_adv}
\alias{write_adv}
\title{Write an MT3DMS advection package file}
\usage{
write_adv(adv, file, IPRN = -1)
}
\arguments{
\item{adv}{an \code{\link{RMT3DMS}} adv object}

\item{file}{filename to write to; typically '*.adv'}

\item{IPRN}{format code for printing arrays in the listing file; defaults to -1 (no printing)}
}
\value{
\code{NULL}
}
\description{
Write an MT3DMS advection package file
}

