% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_btn.R
\name{write_btn}
\alias{write_btn}
\title{Write an MT3DMS basic transport package file}
\usage{
write_btn(btn, file, IPRN = -1)
}
\arguments{
\item{btn}{an \code{\link{RMT3DMS}} btn object}

\item{file}{filename to write to; typically '*.btn'}

\item{IPRN}{format code for printing arrays in the listing file; defaults to -1 (no printing)}
}
\value{
\code{NULL}
}
\description{
Write an MT3DMS basic transport package file
}

