% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rct.R
\name{write_rct}
\alias{write_rct}
\title{Write an MT3DMS file}
\usage{
write_rct(rct, file = {    
  cat("Please select rct file to overwrite or provide new filename ...\\n")    
  file.choose() }, btn, iprn = -1)
}
\arguments{
\item{rct}{an \code{\link{RMT3DMS}} rct object}

\item{file}{filename to write to; typically '*.rct'}

\item{btn}{an \code{\link{RMT3DMS}} btn object}

\item{iprn}{format code for printing arrays in the listing file; defaults to -1 (no printing)}
}
\value{
\code{NULL}
}
\description{
Write an MT3DMS file
}

