% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ssm.R
\name{write_ssm}
\alias{write_ssm}
\title{Write an MT3DMS file}
\usage{
write_ssm(ssm, file, btn, IPRN = -1)
}
\arguments{
\item{ssm}{an \code{\link{RMT3DMS}} ssm object}

\item{file}{filename to write to; typically '*.ssm'}

\item{btn}{an \code{\link{RMT3DMS}} btn object}

\item{IPRN}{format code for printing arrays in the listing file; defaults to -1 (no printing)}
}
\value{
\code{NULL}
}
\description{
Write an MT3DMS file
}

