% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cell_centers.btn.R
\name{cell_centers.btn}
\alias{cell_centers.btn}
\title{Get cell center z coordinates from a btn object}
\usage{
\method{cell_centers}{btn}(btn)
}
\arguments{
\item{btn}{btn object}
}
\value{
3d array with cell center z coordinates
}
\description{
Get cell center z coordinates from a btn object
}

