% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.mt3dms_2d_array.R
\name{plot.mt3dms_2d_array}
\alias{plot.mt3dms_2d_array}
\title{Plot a MODFLOW 2D array}
\usage{
\method{plot}{mt3dms_2d_array}(mt3dms_2d_array, btn, mask = btn$ICBUND[, , 1],
  colour_palette = rev_rainbow, zlim = range(mt3dms_2d_array[which(mask !=
  0)], finite = TRUE), levels = pretty(zlim, nlevels), nlevels = 7,
  type = "fill", add = FALSE, plot3d = FALSE, height_exaggeration = 100,
  binwidth = 1, label = TRUE, prj = NULL, target_CRS = NULL,
  alpha = 1, height = NULL)
}
\arguments{
\item{mt3dms_2d_array}{an object of class mt3dms_2d_array, or a 2D matrix}

\item{btn}{basic transport file object}

\item{mask}{a 2D array with 0 or F indicating inactive cells; defaults to the first layer of btn$ICBUND}

\item{colour_palette}{a colour palette for imaging the array values}

\item{zlim}{vector of minimum and maximum value for the colour scale}

\item{levels}{levels to indicate on the colour scale; defaults to nlevels pretty breakpoints}

\item{nlevels}{number of levels for the colour scale; defaults to 7}

\item{type}{plot type: 'fill' (default), 'grid' or 'contour'}

\item{add}{logical; if TRUE, provide ggplot2 layers instead of object, or add 3D plot to existing rgl device; defaults to FALSE}

\item{plot3d}{logical; should a 3D plot be made}

\item{height_exaggeration}{height exaggeration for 3D plot; optional}

\item{binwidth}{binwidth for contour plot; defaults to 1/20 of zlim}

\item{label}{logical; should labels be added to contour plot}

\item{prj}{projection file object}

\item{target_CRS}{coordinate reference system for the plot}

\item{alpha}{transparency value; defaults to 1}

\item{height}{2D array for specifying the 3D plot z coordinate}
}
\value{
ggplot2 object or layer; if plot3D is TRUE, nothing is returned and the plot is made directly
}
\description{
\code{plot.mt3dms_2d_array} plots a MODFLOW 2D array.
}

