% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.mt3dms_3d_array.R
\name{plot.mt3dms_3d_array}
\alias{plot.mt3dms_3d_array}
\title{Plot a layer of a MT3DMS 3D array}
\usage{
\method{plot}{mt3dms_3d_array}(mt3dms_3d_array, i = NULL, j = NULL,
  k = NULL, btn, mask = btn$ICBUND,
  zlim = range(mt3dms_3d_array[ifelse0(is.null(i),
  c(1:dim(mt3dms_3d_array)[1]), i), ifelse0(is.null(j),
  c(1:dim(mt3dms_3d_array)[2]), j), ifelse0(is.null(k),
  c(1:dim(mt3dms_3d_array)[3]), k)], finite = TRUE),
  color.palette = rev_rainbow, nlevels = 7, ...)
}
\arguments{
\item{mt3dms_3d_array}{an object of class mt3dms_3d_array, or a 3D array}

\item{i}{row number to plot}

\item{j}{column number to plot}

\item{k}{layer number to plot}

\item{btn}{basic transport file object}

\item{mask}{a 3D array with 0 or F indicating inactive cells; defaults to btn$ICBUND}

\item{zlim}{vector of minimum and maximum value for the colour scale}

\item{color.palette}{A color palette for imaging the parameter values}

\item{nlevels}{number of levels for the colour scale; defaults to 7}

\item{...}{arguments provided to plot.mt3dms_2d_array}
}
\value{
ggplot2 object or layer; if plot3D is TRUE, nothing is returned and the plot is made directly
}
\description{
\code{plot.mt3dms_3d_array} plots a 2D section through a MT3DMS 3D array.
}

