% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.mt3dms_3d_array}
\alias{plot.mt3dms_3d_array}
\title{Plot a layer of a MODFLOW 3D array}
\usage{
\method{plot}{mt3dms_3d_array}(mt3dms_3d_array, layer = 1, btn,
  ibound = mt3dms_3d_array * 0 + 1, zlim = range(mt3dms_3d_array[, ,
  layer][which(ibound != 0)], finite = TRUE), ...)
}
\arguments{
\item{mt3dms_3d_array}{An object of class mt3dms_3d_array, or a 3D array}

\item{layer}{The number of the layer to plot}

\item{ibound}{A 3D ibound array with 1 or TRUE indicating active cells, and 0 or F indicating inactive cells}

\item{zlim}{}

\item{color.palette}{A color palette for imaging the parameter values}

\item{levels}{}

\item{nlevels}{}

\item{main}{}
}
\value{
None
}
\description{
\code{plot.mt3dms_3d_array} plots a layer of a MODFLOW 3D array.
}

