% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_adv.R
\name{read_adv}
\alias{read_adv}
\title{Read an MT3DMS advection package file}
\usage{
read_adv(file)
}
\arguments{
\item{file}{filename; typically '*.adv'}
}
\value{
object of class adv
}
\description{
\code{read_adv} reads in an MT3DMS advection package file and returns it as an \code{\link{RMT3DMS}} adv object.
}

