% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_dsp}
\alias{read_dsp}
\title{Read an MT3DMS dispersion package file}
\usage{
read_dsp(file, btn)
}
\arguments{
\item{file}{Filename; typically *.dsp}
}
\value{
Object of class dsp
}
\description{
\code{read.dsp} reads in an MT3DMS dispersion package file and returns it as an \code{\link{RMT3DMS}} dsp object.
}

