% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_gcg}
\alias{read_gcg}
\title{Read an MT3DMS generalized conjugate gradient solver package file}
\usage{
read_gcg(file)
}
\arguments{
\item{file}{Filename; typically *.gcg}
}
\value{
Object of class gcg
}
\description{
\code{read.gcg} reads in an MT3DMS generalized conjugate gradient solver package file and returns it as an \code{\link{RMT3DMS}} gcg object.
}

