% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_mas.R
\name{read_mas}
\alias{read_mas}
\title{Read an MT3DMS mass balance summary file}
\usage{
read_mas(file)
}
\arguments{
\item{file}{filename; typically '*.mas'}
}
\value{
object of class mas
}
\description{
\code{read_mas} reads in an MT3DMS mass balance summary file and returns it as an \code{\link{RMT3DMS}} mas object.
}

