% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmt_convert_dis_to_btn}
\alias{rmt_convert_dis_to_btn}
\title{Convert RMT3DMS dis to RMT3DMS btn object}
\usage{
rmt_convert_dis_to_btn(
  dis,
  nper = dis$nper,
  perlen = dis$perlen,
  tunit = NULL,
  lunit = NULL,
  ...
)
}
\arguments{
\item{dis}{\code{RMODFLOW} dis object}

\item{...}{additional arguments passed to \code{rmt_create_btn}}
}
\value{
\code{RMT3DMS} btn object
}
\description{
Convert RMT3DMS dis to RMT3DMS btn object
}
\details{
conversion works best when number of stress-periods is equal in the dis and btn objects.
}
\examples{
library(RMODFLOW)
dis <- rmf_read_dis(rmf_example_file('example-model.dis'))
rmt_convert_dis_to_btn(dis)
rmt_convert_dis_to_btn(dis, nper = 2, perlen = c(1000, 500), tunit = 'd', ncomp = 2)

}
\seealso{
\code{\link{rmt_convert_btn_to_dis}}, \code{\link{rmt_create_btn}}
}
