% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{rmt_install}
\alias{rmt_install}
\title{Install external codes}
\usage{
rmt_install(code = "all", overwrite = NULL)
}
\arguments{
\item{code}{Character vector with the codes to install, or `"all"` (default).}

\item{overwrite}{Logical. Overwrite when the code is already installed? If
`NULL` (default), the user is asked what to do in an interactive session.
An error message is issued otherwise.}
}
\description{
This function installs external codes that are used by [RMT3DMS].
}
\details{
## Supported software
[MT3DMS v5.3](https://hydro.geo.ua.edu/) and [MT3D-USGS](https://www.usgs.gov/software/mt3d-usgs-groundwater-solute-transport-simulator-modflow). 
The zip files with windows binaries hosted at these website
are downloaded and extracted in the installation directory. The main
folder names are modified in order to have more consistency.

## Installation location
The default installation location is `file.path(system.file(package =
"RMT3DMS"), "code")`, but it can be altered by setting option
`RMT3DMS.path`.
}
\examples{
rmt_install() # Install all codes.
rmt_install("MT3D-USGS", overwrite = TRUE) # Install MT3D-USGS.
rmt_install("MT3DMS", overwrite = TRUE) # Install MT3DMS.
}
