% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmt_plot.ucn}
\alias{rmt_plot.ucn}
\title{Plot a MT3D ucn file object}
\usage{
\method{rmt_plot}{ucn}(
  ucn,
  btn,
  i = NULL,
  j = NULL,
  k = NULL,
  l = NULL,
  kper = NULL,
  kstp = NULL,
  ...
)
}
\arguments{
\item{ucn}{\code{RMT3DMS} ucn object as returned from \code{rmt_read_ucn}}

\item{btn}{\code{RMT3DMS} btn object}

\item{i}{row number to plot}

\item{j}{column number to plot}

\item{k}{layer number to plot}

\item{l}{time step number to plot; defaults to plotting the final time step.}

\item{kper}{integer specifying the stress-period. Use in conjunction with kstp. See details.}

\item{kstp}{integer specifying the time step of kper. Use in conjunction with kper. See details.}

\item{...}{additional parameters passed to \code{\link{rmt_plot.rmf_4d_array}}}
}
\value{
ggplot2 object or layer
}
\description{
Plot a MT3D ucn file object
}
\details{
There are two ways to specify which time step to plot. The \code{l} argument can be specified which represents the total time step number. 
 The other option is to specify both \code{kper} & \code{kstp} which specify the stress-period and corresponding time step in that stress-period.
 A negative \code{kstp} will plot the final time step of the stress-period.
 
 If no output is written for the specified time step, as controlled by the btn object, an error is thrown.
}
