% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dsp.R
\name{write_dsp}
\alias{write_dsp}
\title{Write an MT3DMS dispersion package file}
\usage{
write_dsp(dsp, file, btn, IPRN = -1)
}
\arguments{
\item{dsp}{an \code{\link{RMT3DMS}} dsp object}

\item{file}{filename to write to; typically '*.dsp'}

\item{btn}{an \code{\link{RMT3DMS}} btn object}

\item{IPRN}{format code for printing arrays in the listing file; defaults to -1 (no printing)}
}
\value{
\code{NULL}
}
\description{
Write an MT3DMS dispersion package file
}

